+++ src/components/Tooltip/index.ts
@@ -1,2 +1,7 @@
 export { Tooltip } from "./Tooltip";
 export type { TooltipProps, TooltipContent } from "./Tooltip";
+export { PopoverTooltip } from "./PopoverTooltip";
+export type {
+  PopoverTooltipProps,
+  PopoverTooltipContent,
+} from "./PopoverTooltip";